Connected Main Page and Added Styling
I've connected the main page to the stop page so the links for each stop
are now working.  I also added the stylesheet to the stop page and, in
order to test everything, added all of the code currently in the
"rentals" folder to my leda account.  I made the changes to the main
page's code on my leda account, so the main page in the respository
hasn't been changed yet.  I can make the changes if you guys want
though.